Changed id type in FlightRepository

+++ src/main/java/com/zakat/air_tickets/repository/FlightRepository.java
@@ -8,7 +8,7 @@ import java.sql.Timestamp;
 import java.util.List;
 import java.util.Set;
 
-public interface FlightRepository extends JpaRepository<Flight, Integer> {
+public interface FlightRepository extends JpaRepository<Flight, Long> {
     @Query(
             "SELECT f.departureCity FROM Flight f"
     )
